+++ src/pages/scientificServices/pipelines/tabs/history/details/sections/timeline/PipelineRunTimeline.tsx
@@ -2,8 +2,9 @@ import { Icon } from '@terra-ui-packages/components';
 import React from 'react';
 import { PipelineRunResponse } from 'src/libs/ajax/teaspoons/teaspoons-models';
 import colors from 'src/libs/colors';
+import { calculateTimelineEvents } from 'src/pages/scientificServices/pipelines/tabs/history/details/sections/timeline/pipeline-timeline-utils';
 
-import { calculateRunDuration, calculateTimelineEvents, PipelineRunTimelineEvent } from './PipelineRunTimelineEvent';
+import { PipelineRunTimelineEvent } from './PipelineRunTimelineEvent';
 
 interface PipelineRunTimelineProps {
   pipelineRunResult: PipelineRunResponse;
@@ -52,4 +53,28 @@ export const PipelineRunTimeline = ({ pipelineRunResult }: PipelineRunTimelinePr
       ))}
     </div>
   );
+};
+
+const calculateRunDuration = (pipelineRunResult: PipelineRunResponse): string => {
+  if (!pipelineRunResult.jobReport.submitted || !pipelineRunResult.jobReport.completed) {
+    return 'N/A';
+  }
+
+  const durationMs =
+    new Date(pipelineRunResult.jobReport.completed).getTime() -
+    new Date(pipelineRunResult.jobReport.submitted).getTime();
+
+  if (durationMs === 0) {
+    return 'In progress';
+  }
+
+  const totalSeconds = Math.floor(durationMs / 1000);
+  const hours = Math.floor(totalSeconds / 3600);
+  const minutes = Math.floor((totalSeconds % 3600) / 60);
+  const seconds = totalSeconds % 60;
+
+  if (hours > 0) {
+    return `${hours}h ${minutes}m ${seconds}s`;
+  }
+  return `${minutes}m ${seconds}s`;
 };
